PART 2 OUT YEEHAW DEBUG TIME

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: vtarasov <vtarasov@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/16 14:36:55 by vtarasov          #+#    #+#             */
-/*   Updated: 2026/06/17 20:40:29 by vtarasov         ###   ########.fr       */
+/*   Updated: 2026/06/24 20:23:53 by vtarasov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	a;
+	size_t	a;
 
 	a = 0;
 	while (a < n)
